L'application stock désormais correctement le nombre de like que possède un restaurant même lorsque l'on change de fragments ou d'activités. Correction d'un problème sur les notifications qui se lançait juste après avoir choisi un restaurant

+++ app/src/main/java/com/example/go4lunch/api/RestaurantFirebaseHelper.java
@@ -14,7 +14,7 @@ public class RestaurantFirebaseHelper {
         return FirebaseFirestore.getInstance().collection(COLLECTION_NAME);
     }
     // --- CREATE ---
-    public static Task<Void> createRestaurantFirebase(String id, int likeNumber) {
+    public static Task<Void> createRestaurantFirebase(String id, long likeNumber) {
         MyRestaurantFirebase  myRestaurantFirebase= new MyRestaurantFirebase(id, likeNumber);
         return RestaurantFirebaseHelper.getRestaurantFirebaseCollection().document(id)
                 .set(myRestaurantFirebase);
@@ -26,7 +26,7 @@ public class RestaurantFirebaseHelper {
     }
 
     // --- UPDATE ---
-    public static Task<Void> updateRestaurantFirebase(String id, int likeNumber) {
+    public static Task<Void> updateRestaurantFirebase(String id, long likeNumber) {
         return RestaurantFirebaseHelper.getRestaurantFirebaseCollection().document(id)
                 .update("likeNumber", likeNumber);
     }
